Se actualizaron el tipo de dato en algunas variables

+++ app/src/main/java/com/example/easysoccer1/ui/view/GoalsFragment.kt
@@ -80,9 +80,9 @@ class GoalsFragment : Fragment() {
                     if (size == "5vs5") {
                         price5vs5?.let { it1 ->
                             Goals(
-                                number = binding.inputTextNumberGoal.text.toString(),
+                                number = binding.inputTextNumberGoal.text.toString().toInt(),
                                 size = size,
-                                price = it1,
+                                price = it1.toInt(),
                                 available = "Disponible",
                                 hour = "",
                                 date = ""
@@ -95,9 +95,9 @@ class GoalsFragment : Fragment() {
                     } else {
                         price8vs8?.let { it1 ->
                             Goals(
-                                number = binding.inputTextNumberGoal.text.toString(),
+                                number = binding.inputTextNumberGoal.text.toString().toInt(),
                                 size = size,
-                                price = it1,
+                                price = it1.toInt(),
                                 available = "Disponible",
                                 hour = "",
                                 date = ""
